rearranged builder code
