--- src/builder.spec.mts
@@ -1,9 +1,0 @@
-// builder.spec.mts
-
-import './builder';
-
-describe('test builder', () => {
-  it('should build', () => {
-    expect(true).toBe(true);
-  });
-});
